+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+##### 17.06.2025
+
+> Update: Rates table standard rates updates
+> Update: Owner's report entry catgeory label update
+
+##### 31.05.2025
+
+> Feature: Condensed pagination support added
+
 ##### 28.05.2025
 
 > Feature: Banks list selection and verification added